refactor(swap): get rid of mod files
